MAINT: fix a bunch of minor issues
