+++ odl/discr/grid.py
@@ -523,7 +523,7 @@ class RectGrid(Set):
     def __hash__(self):
         """Return ``hash(self)``."""
         # TODO: update with #841
-        coord_vec_str = tuple(cv.tostring() for cv in self.coord_vectors)
+        coord_vec_str = tuple(cv.tobytes() for cv in self.coord_vectors)
         return hash((type(self), coord_vec_str))
 
     def approx_contains(self, other, atol):
