+++ odl/set/space.py
@@ -52,11 +52,6 @@ class LinearSpace(Set):
         """Scalar field of numbers for this vector space."""
         return self.__field
 
-    @property
-    def has_field(self):
-        """``True`` if `field` is not ``None``, otherwise ``False``."""
-        return self.field is not None
-
     def element(self, inp=None, **kwargs):
         """Create a `LinearSpaceElement` from ``inp`` or from scratch.
 
@@ -203,7 +198,7 @@ class LinearSpace(Set):
 
             ``x = x * (1 + 2 + 3.14)``.
         """
-        if not self.has_field:
+        if self.field is None:
             raise TypeError('`lincomb` cannot be used with `field=None`')
 
         if out is None:
@@ -290,7 +285,7 @@ class LinearSpace(Set):
         inner : `LinearSpace.field` element
             Inner product of ``x1`` and ``x2``.
         """
-        if not self.has_field:
+        if self.field is None:
             raise TypeError('`inner` cannot be used with `field=None`')
         if x1 not in self:
             raise LinearSpaceTypeError('`x1` {!r} is not an element of '
@@ -497,7 +492,9 @@ class LinearSpaceElement(object):
     # Convenience methods
     def __iadd__(self, other):
         """Implement ``self += other``."""
-        if other in self.space:
+        if self.space.field is None:
+            return NotImplemented
+        elif other in self.space:
             return self.space.lincomb(1, self, 1, other, out=self)
         elif isinstance(other, LinearSpaceElement):
             # We do not `return NotImplemented` here since we don't want a
@@ -527,6 +524,8 @@ class LinearSpaceElement(object):
         # Instead of using __iadd__ we duplicate code here for performance
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__radd__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space:
             tmp = self.space.element()
             return self.space.lincomb(1, self, 1, other, out=tmp)
@@ -556,7 +555,9 @@ class LinearSpaceElement(object):
 
     def __isub__(self, other):
         """Implement ``self -= other``."""
-        if other in self.space:
+        if self.space.field is None:
+            return NotImplemented
+        elif other in self.space:
             return self.space.lincomb(1, self, -1, other, out=self)
         elif isinstance(other, LinearSpaceElement):
             # We do not `return NotImplemented` here since we don't want a
@@ -564,6 +565,8 @@ class LinearSpaceElement(object):
             # `self = self - other` which does not modify self.
             raise TypeError('cannot subtract {!r} and {!r} in-place'
                             ''.format(self, other))
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space.field:
             one = getattr(self.space, 'one', None)
             if one is None:
@@ -585,6 +588,8 @@ class LinearSpaceElement(object):
         # Instead of using __isub__ we duplicate code here for performance
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__rsub__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space:
             tmp = self.space.element()
             return self.space.lincomb(1, self, -1, other, out=tmp)
@@ -609,6 +614,8 @@ class LinearSpaceElement(object):
         """Return ``other - self``."""
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__sub__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space:
             tmp = self.space.element()
             return self.space.lincomb(1, other, -1, self, out=tmp)
@@ -633,7 +640,9 @@ class LinearSpaceElement(object):
 
     def __imul__(self, other):
         """Implement ``self *= other``."""
-        if other in self.space.field:
+        if self.space.field is None:
+            return NotImplemented
+        elif other in self.space.field:
             return self.space.lincomb(other, self, out=self)
         elif other in self.space:
             return self.space.multiply(other, self, out=self)
@@ -657,6 +666,8 @@ class LinearSpaceElement(object):
         # Instead of using __imul__ we duplicate code here for performance
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__rmul__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space.field:
             tmp = self.space.element()
             result = self.space.lincomb(other, self, out=tmp)
@@ -683,6 +694,8 @@ class LinearSpaceElement(object):
 
     def __itruediv__(self, other):
         """Implement ``self /= other``."""
+        if self.space.field is None:
+            return NotImplemented
         if other in self.space.field:
             return self.space.lincomb(1.0 / other, self, out=self)
         elif other in self.space:
@@ -708,6 +721,8 @@ class LinearSpaceElement(object):
         """Return ``self / other``."""
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__rtruediv__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space.field:
             tmp = self.space.element()
             return self.space.lincomb(1.0 / other, self, out=tmp)
@@ -730,6 +745,8 @@ class LinearSpaceElement(object):
         """Return ``other / self``."""
         if getattr(other, '__array_priority__', 0) > self.__array_priority__:
             return other.__truediv__(self)
+        elif self.space.field is None:
+            return NotImplemented
         elif other in self.space.field:
             one = getattr(self.space, 'one', None)
             if one is None:
@@ -758,6 +775,8 @@ class LinearSpaceElement(object):
         """Implement ``self ** p``.
 
         This is only defined for integer ``p``."""
+        if self.space.field is None:
+            return NotImplemented
         p, p_in = int(p), p
         if p != p_in:
             raise ValueError('expected integer `p`, got {}'.format(p_in))
@@ -783,12 +802,16 @@ class LinearSpaceElement(object):
 
     def __pow__(self, p):
         """Return ``self ** p``."""
+        if self.space.field is None:
+            return NotImplemented
         tmp = self.copy()
         tmp.__ipow__(p)
         return tmp
 
     def __neg__(self):
         """Return ``-self``."""
+        if self.space.field is None:
+            return NotImplemented
         return (-1) * self
 
     def __pos__(self):
